+++ setup.py
@@ -10,6 +10,7 @@ setup(name='tap-s3-csv',
       classifiers=['Programming Language :: Python :: 3 :: Only'],
       py_modules=['tap_s3_csv'],
       install_requires=[
+          'backoff==1.3.2',
           'boto3==1.4.4',
           'singer-encodings==0.0.3',
           'singer-python==5.1.5',
